requirements.txt
pandas == 2.3.0
plotly_express == 0.4.1
streamlit == 1.45.1
